Fixing int and double

+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_1/Scanner.java
@@ -9,7 +9,7 @@ public class Scanner {
         System.out.println( numberA + numberB );
         System.out.println( numberA - numberB );
         System.out.println( numberA * numberB );
-        System.out.println( numberA / numberB );
+        System.out.println( numberA / (double) numberB );
 
     }
 }
